chore: rename bencher to clarify
